refactor: renomeando componente (padrão ingles)
